requirement.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.29
python-dotenv==1.0.1